Multi rtorrent instances support (#549)
* Handling multiple rtorrent instances

* performance issue addressing

* fix wrong conflict resolution

* performances

* purge services instance every 10 mins

* last fixes

* avoid duplicated reducers

+++ server/models/Users.js
@@ -1,5 +1,6 @@
 'use strict';
 const argon2 = require('argon2');
+const fs = require('fs-extra');
 const Datastore = require('nedb');
 
 const config = require('../../config');
@@ -35,7 +36,14 @@ class Users {
   }
 
   createUser(credentials, callback) {
-    const {password, username} = credentials;
+    const {
+      password,
+      username,
+      host,
+      port,
+      socketPath,
+      isAdmin
+    } = credentials;
 
     if (!this.ready) {
       return callback(null, 'Users database is not ready.');
@@ -45,10 +53,12 @@ class Users {
       return callback(null, 'Username cannot be empty.');
     }
 
+    const socket = socketPath != null && socketPath !== '' && 0 < socketPath.trim().length;
+
     argon2
       .hash(password)
       .then(hash => {
-        this.db.insert({ username, password: hash }, (error, user) => {
+        this.db.insert({ username, password: hash, host, port, socket, socketPath, isAdmin }, (error, user) => {
           if (error) {
             if (error.errorType === 'uniqueViolated') {
               error = 'Username already exists.';
@@ -64,12 +74,25 @@ class Users {
   }
 
   removeUser(username, callback) {
-    this.db.remove({username: username}, {}, (err, numRemoved) => {
+    this.db.findOne({username: username}).exec((err, user) => {
       if (err) {
         return callback(null, err);
       }
 
-      return callback({username: username});
+      // Username not found.
+      if (user == null) {
+        return callback(null, user);
+      }
+
+      this.db.remove({username: username}, {}, (err, numRemoved) => {
+        if (err) {
+          return callback(null, err);
+        }
+
+        fs.removeSync(`${config.dbPath}${user._id}/`);
+
+        return callback({username: username});
+      });
     });
   }
 
@@ -97,6 +120,16 @@ class Users {
 
   lookupUser(credentials, callback) {
     this.db.findOne({username: credentials.username}, (err, user) => {
+      if (err) {
+        return callback(err);
+      }
+
+      return callback(null, user);
+    });
+  }
+
+  fetchById(userId, callback) {
+    this.db.findOne({_id: userId}, (err, user) => {
       if (err) {
         return callback(err);
       }
